backend/requirements.txt
cx_Oracle==8.3.0
faiss_cpu==1.8.0.post1
#faiss_gpu_cu12==1.8.0.2
fitz==0.0.1.dev2
gspread==6.1.2
inputimeout==1.0.4
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.35
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.2
Markdown==3.7
markdownify==0.13.1
oauth2client==4.1.3
opencv_python==4.10.0.84
Pillow==10.4.0
pynput==1.7.7
pytesseract==0.3.10
Requests==2.32.3
scikit_learn==1.5.1
torch==2.3.1
transformers==4.42.4
